[Upd] moved game in singleton + upd methods

+++ singleton/game.py
@@ -17,21 +17,10 @@ class Game:
         else:
             Game.instance = self
 
-        with open("assets/init_db.sql", "r") as sql_file:
-            sql_script = sql_file.read()
-
-        cursor.executescript(sql_script)
-        db.commit()
-
     @connected
     def addWord(self, word: str, user, db, cursor):
         user_id = cursor.execute("SELECT ID FROM Users WHERE discord_id=?", (str(user.id),)).fetchall()
-        if len(user_id) == 0:
-            user_id = self.addUser(user)
-        else:
-            user_id = user_id[0][0]
 
-        print(user_id)
         cursor.execute("INSERT INTO Words (word, creator_id) VALUES (?, ?)", (word, user_id))
         db.commit()
 
@@ -44,19 +33,15 @@ class Game:
 
     @connected
     def delWord(self, word: str, db, cursor):
-        cursor.execute("DELETE FROM Words WHERE word=?", (word,))
-        db.commit()
+        exists = cursor.execute("SELECT COUNT(*) FROM Words WHERE word=?", (word,)).fetchall()[0][0] == 1
+        if exists:
+            cursor.execute("DELETE FROM Words WHERE word=?", (word,))
+            db.commit()
+
+        return exists
 
     def startGame(self):
         pass
-
-    @connected
-    def addUser(self, user, db, cursor):
-        cursor.execute(
-            "INSERT INTO Users (username, discord_id, score) VALUES (?, ?, ?)", (user.name, str(user.id), 0)
-        )
-        db.commit()
-        return cursor.execute("SELECT last_insert_rowid()").fetchall()[0]
 
     def addUserToGame(self, user_id: str, game_id: str):
         pass
